+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Fixed
+- Keep the *last* `Waker` in `Future::poll`, not the *first* one.
 
 
 ## [0.1.2] - 2020-08-11